[MINOR] Migrate RankValue to the package of the common class
