+++ coordinator/src/main/java/org/apache/uniffle/coordinator/LowestIOSampleCostSelectStorageStrategy.java
@@ -33,6 +33,7 @@ import org.apache.hadoop.fs.Path;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import org.apache.uniffle.common.RankValue;
 import org.apache.uniffle.common.RemoteStorageInfo;
 import org.apache.uniffle.common.filesystem.HadoopFilesystemProvider;
 
@@ -133,58 +134,5 @@ public class LowestIOSampleCostSelectStorageStrategy extends AbstractSelectStora
     }
     LOG.warn("No remote storage is available, we will default to the first.");
     return availableRemoteStorageInfo.values().iterator().next();
-  }
-
-  static class RankValue {
-    AtomicLong costTime;
-    AtomicInteger appNum;
-    AtomicBoolean isHealthy;
-
-    RankValue(int appNum) {
-      this.costTime = new AtomicLong(0);
-      this.appNum = new AtomicInteger(appNum);
-      this.isHealthy = new AtomicBoolean(true);
-    }
-
-    RankValue(long costTime, int appNum) {
-      this.costTime = new AtomicLong(costTime);
-      this.appNum = new AtomicInteger(appNum);
-      this.isHealthy = new AtomicBoolean(true);
-    }
-
-    public AtomicLong getCostTime() {
-      return costTime;
-    }
-
-    public AtomicInteger getAppNum() {
-      return appNum;
-    }
-
-    public AtomicBoolean getHealthy() {
-      return isHealthy;
-    }
-
-    public void setCostTime(AtomicLong readAndWriteTime) {
-      this.costTime = readAndWriteTime;
-    }
-
-    public void setAppNum(AtomicInteger appNum) {
-      this.appNum = appNum;
-    }
-
-    public void setHealthy(AtomicBoolean isHealthy) {
-      this.isHealthy = isHealthy;
-      if (!isHealthy.get()) {
-        this.costTime.set(Long.MAX_VALUE);
-      }
-    }
-
-    @Override
-    public String toString() {
-      return "RankValue{"
-          + "costTime=" + costTime
-          + ", appNum=" + appNum
-          + '}';
-    }
   }
 }
